Fix errors and warnings by open-watcom
Moved assembly functions ___save_flags and _restore_flags from
arch/i86/kernel/irq.c to arch/i86/kernel/irqtab.c. Now, irq.c
is an all C file.

Updated arch/i86/sibo/irqtab .c to stay in sync with
arch/i86/kernel/irqtab.c.

Removed garbage from irqtab.c, include/linuxmt/sched.h,
fs/minix/inode.c and net/socket.c

Small changes in arch/i86/kernel/signal.c, fs/minix/dir.c,
fs/minix/file.c, fs/readdir.c, include/arch/irq.h,
include/linuxmt/fs.h, include/linuxmt/minix_fs.h,
include/linuxmt/mm.h, include/linuxmt/tcpdev.h and
net/ipv4/af_inet.c to remove warnings.

Code size was reduced by 16 bytes and data size increased
by 16 bytes.

+++ elks/fs/minix/dir.c
@@ -16,7 +16,7 @@
 
 #include <arch/segment.h>
 
-static int minix_dir_read(struct inode *inode, struct file *filp, char *buf, int count)
+static size_t minix_dir_read(struct inode *inode, struct file *filp, char *buf, int count)
 {
     return -EISDIR;
 }
